move graphrag.index.storage to graphrag.storage
