+++ graphrag/index/cache/json_pipeline_cache.py
@@ -7,7 +7,7 @@ import json
 from typing import Any
 
 from graphrag.index.cache.pipeline_cache import PipelineCache
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 class JsonPipelineCache(PipelineCache):
