+++ graphrag/index/input/text.py
@@ -11,9 +11,9 @@ from typing import Any
 import pandas as pd
 
 from graphrag.index.config.input import PipelineInputConfig
-from graphrag.index.storage.pipeline_storage import PipelineStorage
 from graphrag.index.utils.hashing import gen_md5_hash
 from graphrag.logging.base import ProgressReporter
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 DEFAULT_FILE_PATTERN = re.compile(
     r".*[\\/](?P<source>[^\\/]+)[\\/](?P<year>\d{4})-(?P<month>\d{2})-(?P<day>\d{2})_(?P<author>[^_]+)_\d+\.txt"
