+++ graphrag/index/operations/snapshot.py
@@ -5,7 +5,7 @@
 
 import pandas as pd
 
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 async def snapshot(
