+++ graphrag/index/operations/snapshot_graphml.py
@@ -5,7 +5,7 @@
 
 import networkx as nx
 
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 async def snapshot_graphml(
