+++ graphrag/index/operations/snapshot_rows.py
@@ -9,7 +9,7 @@ from typing import Any
 
 import pandas as pd
 
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @dataclass
