+++ graphrag/index/update/incremental_index.py
@@ -13,7 +13,6 @@ from graphrag.index.cache.pipeline_cache import PipelineCache
 from graphrag.index.config.pipeline import PipelineConfig
 from graphrag.index.flows.generate_text_embeddings import generate_text_embeddings
 from graphrag.index.run.workflow import _find_workflow_config
-from graphrag.index.storage.pipeline_storage import PipelineStorage
 from graphrag.index.update.communities import (
     _merge_and_resolve_nodes,
     _update_and_merge_communities,
@@ -25,6 +24,7 @@ from graphrag.index.update.entities import (
 )
 from graphrag.index.update.relationships import _update_and_merge_relationships
 from graphrag.logging.print_progress import ProgressReporter
+from graphrag.storage.pipeline_storage import PipelineStorage
 from graphrag.utils.storage import _load_table_from_storage
