+++ graphrag/index/workflows/v1/subflows/create_base_entity_graph.py
@@ -18,7 +18,7 @@ from graphrag.index.cache.pipeline_cache import PipelineCache
 from graphrag.index.flows.create_base_entity_graph import (
     create_base_entity_graph as create_base_entity_graph_flow,
 )
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @verb(
