+++ graphrag/index/workflows/v1/subflows/create_final_covariates.py
@@ -17,7 +17,7 @@ from graphrag.index.cache.pipeline_cache import PipelineCache
 from graphrag.index.flows.create_final_covariates import (
     create_final_covariates as create_final_covariates_flow,
 )
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @verb(name="create_final_covariates", treats_input_tables_as_immutable=True)
