+++ graphrag/index/workflows/v1/subflows/create_final_documents.py
@@ -16,7 +16,7 @@ from datashaper.table_store.types import VerbResult, create_verb_result
 from graphrag.index.flows.create_final_documents import (
     create_final_documents as create_final_documents_flow,
 )
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @verb(
