+++ graphrag/index/workflows/v1/subflows/create_final_entities.py
@@ -15,7 +15,7 @@ from datashaper.table_store.types import VerbResult, create_verb_result
 from graphrag.index.flows.create_final_entities import (
     create_final_entities as create_final_entities_flow,
 )
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @verb(
