+++ graphrag/index/workflows/v1/subflows/create_final_nodes.py
@@ -15,7 +15,7 @@ from datashaper.table_store.types import VerbResult, create_verb_result
 from graphrag.index.flows.create_final_nodes import (
     create_final_nodes as create_final_nodes_flow,
 )
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 
 @verb(name="create_final_nodes", treats_input_tables_as_immutable=True)
