+++ graphrag/storage/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""The Indexing Engine storage package root."""
+"""Storage package root."""
