+++ graphrag/storage/file_pipeline_storage.py
@@ -16,8 +16,8 @@ from aiofiles.os import remove
 from aiofiles.ospath import exists
 from datashaper import Progress
 
-from graphrag.index.storage.pipeline_storage import PipelineStorage
 from graphrag.logging.base import ProgressReporter
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 log = logging.getLogger(__name__)
