+++ graphrag/utils/storage.py
@@ -8,7 +8,7 @@ from io import BytesIO
 
 import pandas as pd
 
-from graphrag.index.storage.pipeline_storage import PipelineStorage
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 log = logging.getLogger(__name__)
