+++ tests/unit/indexing/cache/test_file_pipeline_cache.py
@@ -5,7 +5,7 @@ import os
 import unittest
 
 from graphrag.index.cache.json_pipeline_cache import JsonPipelineCache
-from graphrag.index.storage.file_pipeline_storage import (
+from graphrag.storage.file_pipeline_storage import (
     FilePipelineStorage,
 )
